Add BaseNNEstimator parent class for NNMixture and NF
